+++ daemon/oci_linux.go
@@ -15,12 +15,12 @@ import (
 	"github.com/containerd/log"
 	dconfig "github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/container"
+	"github.com/docker/docker/daemon/internal/rootless/mountopts"
+	"github.com/docker/docker/daemon/internal/rootless/specconv"
 	"github.com/docker/docker/daemon/pkg/oci"
 	"github.com/docker/docker/daemon/pkg/oci/caps"
 	volumemounts "github.com/docker/docker/daemon/volume/mounts"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/internal/rootless/mountopts"
-	"github.com/docker/docker/internal/rootless/specconv"
 	containertypes "github.com/moby/moby/api/types/container"
 	"github.com/moby/sys/mount"
 	"github.com/moby/sys/mountinfo"
